Reworked some SelfDespawn logic into a new class, SelfWorldBoundsDespawn
