+++ Assets/Scripts/SelfWorldBoundsDespawn.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: da6900ffe20eb584c89dbe88241958e3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
